+++ resources/newssite/news-nuxt/app.vue
@@ -16,10 +16,16 @@ if (process.client) {
 }
 </script>
 
+<script setup>
+import { provideLocale } from "./composables/provide-locale";
+import { scrollOnNavigation } from "./composables/scroll-behavior";
+
+provideLocale();
+scrollOnNavigation();
+</script>
+
 <template>
-    <DataContext>
-        <NuxtLayout>
-            <NuxtPage />
-        </NuxtLayout>
-    </DataContext>
+    <Layout>
+        <NuxtPage />
+    </Layout>
 </template>
