+++ resources/newssite/news-nuxt/nuxt.config.ts
@@ -6,7 +6,7 @@ import { defineNuxtConfig } from "nuxt/config";
 export default defineNuxtConfig({
     ssr: false,
     css: ["news-site-css/dist/variables.css", "news-site-css/dist/global.css", "news-site-css/dist/a11y.css", "news-site-css/dist/icons.css", "news-site-css/dist/text.css"],
-    components: ["~/components", "~/components/assets", "~/components/atoms", "~/components/molecules", "~/context"],
+    components: ["~/components", "~/components/assets", "~/components/atoms", "~/components/molecules", "~/components/organisms"],
     router: {
         options: {
             hashMode: true,
